Let logging handle pdf responses
Because:
When fetching PDF's with HTTPEx I encountered the following error:

```
[error] ** (ArgumentError) argument error
    (stdlib 6.2) io.erl:203: :io.put_chars(:standard_io, [<<27, 91, 52, 109, 72, 84, 84, 80, 32, 114, 101, 115, 112, 111, 110, 115, 101, 58, 27, 91, 50, 52, 109, 10, 10, 27, 91, 49, 109, 67, 108, 105, 101, 110, 116, 58, 27, 91, 50, 50, 109, 32, 27, 91, 51, 109, 58, 102, 105, ...>>, 10])
    (http_ex 0.1.0) lib/http_ex/logging.ex:43: HTTPEx.Logging.log/1
    (http_ex 0.1.0) lib/http_ex/logging.ex:50: HTTPEx.Logging.log/1
```

I was able to follow this back to the summary of the response, with
Michaël we looked at knowing upfront we had to do this but couldn't find
it. If there's a better way to know how to make a string out of it,
idea's are welcome.

We've now implemented a rescue for this call which uses the original
response or else tries to encode this.

Will add a test for this in a follow up.