Export logging
In the use case of 2 apps we have where one is only to send external
requests and the other to keep state, we want to log the requests we've
done in the app where we keep state (MyWuunder) so that support,
implementation has easier access to the logs in communication with the
external services.

To do this we let the external services app define a function in how
that app wants to send the requests and where to. From HTTPEx point whe
just call this function when it's available, with the logging info.