Validate the options for expect_request!
Because:
When you enter `expected_body: {some_json, :json},` and the tests
succeed you might think you've tested the proper json is send. But
because this option isn't used that's not the case. So to guard against
that we're now checking if the option that's given will actually be
used, rather then ignore it the test will now raise.